+++ webapp/src/main/java/com/infoshare/fourfan/repository/ProductRepositoryBean.java
@@ -3,20 +3,19 @@ package com.infoshare.fourfan.repository;
 import com.infoshare.fourfan.domain.datatypes.Product;
 import com.infoshare.fourfan.domain.datatypes.ProductList;
 import com.infoshare.fourfan.domain.datatypes.Shop;
+import com.infoshare.fourfan.service.AdminService;
 import com.infoshare.fourfan.service.JsonService;
 
 import javax.ejb.Stateless;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
+import java.io.IOException;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Stateless
 public class ProductRepositoryBean implements ProductRepository {
 
     @Override
-    public Optional<Product> findProductById(Long id) {
+    public Optional<Product> findProductById(Integer id) {
         return findAllJson().stream()
                 .filter(product -> product.getId().equals(id)).findFirst();
     }
